[QUAL] Split the static base cache between the cache manager and the cache itself
